+++ src/utils/AuthError.ts
@@ -1,0 +1,6 @@
+export default class AuthError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'AuthError';
+  }
+}